94: CR changes- moved TopicPartition to Commons package
