+++ src/main/java/commons/TopicPartition.java
@@ -1,4 +1,4 @@
-package producer;
+package commons;
 
 import java.util.Objects;
